backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.6
python-dotenv==1.0.0
minio==7.1.2
psycopg2-binary==2.9.10
redis==4.6.0
rq==1.11.0
pdfminer.six==20231228
python-pptx==0.6.21
python-docx==0.8.11
requests==2.31.0
sentence-transformers==2.2.2
# pgvector optional for local installs; skip strict pin to avoid platform issues
neo4j==5.22.0
pytest==7.4.4
# Updated httpx for FastAPI 0.115+
httpx==0.27.2
PyYAML==6.0.2
# LangChain for GraphRAG
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.5
langchain-experimental==0.3.3
